% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coles2019}
\alias{coles2019}
\title{Studies on the Facial Feedback Literature}
\format{
A dataset with 286 rows and 13 variables.
\itemize{
\item \code{study_id}: Unique id for study
\item \code{es_id}: Unique id for effect size
\item \code{yi}: Effect size in Cohen's d
\item \code{vi}: Variance of effect size (SE^2)
\item \code{title}: Title of publication
\item \code{year}: Year of publication
\item \code{file_drawer}: Publication status
\item \code{prop_women}: Proportion of study that identified as women
\item \code{video_know}: Yes: Participants were told they were going to be recorded or the methodology stated that a video camera was placed within participant view.
No" Methodology stated that participants were unaware of video recording, that the video camera was hidden, or that there was not a video camera
\item \code{stim}: Type of stimuli
\item \code{proc}: Type of facial feedback manipulation
\item \code{proc_aware}: Whether participants were aware of the facial feedback manipulation
\item \code{w_v_b}: Whether the study used a between- or within-subjects design
}
}
\source{
\href{https://osf.io/v8kxb/}{https://osf.io/v8kxb/}
}
\usage{
data(coles2019)
}
\description{
Results from 138 studies, including 274] effect sizes (Cohen's d), on the facial feedback hypothesis \insertCite{coles2019meta}{psymetadata}.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
