% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gamble2019}
\alias{gamble2019}
\title{Meta-analytic data collected from}
\format{
* \code{study_id}: Unique id for study
* \code{samp_id}: Unique id for each sample
* \code{es_id}: Unique id for effect size
* \code{authors}: Authors of study
* \code{yi}: Effect size in r
* \code{vi}: Variance of effect size
* \code{ni}: Sample size of study
* \code{sex}: Proportion of study that was female
* \code{age}: Mean age of participants
* \code{dep_status}: Clinical status of depression
* \code{comorbid_anx}: Whehter comorbid with anxiety
* \code{emo_val}: Emotional valence of simulations
* \code{macro_micro}: Macro vs. micro specificity
* \code{cue_type}: Cue type
* \code{spec_rated}: Self- vs. researcher-rated specificity
* \code{dep_rated}: Self- vs. researcher-rated depression
* \code{cat_dim}: Categorical vs. dimensional designs
* \code{quality}: Study quality rating
* \code{published}: Published or not
* \code{mode}: Mode or prospection
}
\source{
\href{https://osf.io/5wjb2/}{Open Science Framework}
}
\usage{
data(gamble2019)
}
\description{
To be filled in.. \insertNoCite{gamble2019specificity}{psychmetadata}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
