% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lowe2020}
\alias{lowe2020}
\title{Studies on the advantage of bilingualism in children: a meta-analytic review}
\format{
A data frame with 1194 rows and 20 variables:
\itemize{
\item \code{pub_year}: year of publication
\item \code{pub_type}: publication type
\item \code{es_id}: effect size id
\item \code{study_id}: study id
\item \code{yi}: effect size (Hedge's g)
\item \code{vi}: sampling variance (SE^2)
\item \code{subsample}: coding for independent subsamples within studies
\item \code{participants}: unique id for participant pairs
\item \code{clusters}: unique id for participant clusters
\item \code{lab_group}: unique id for research group
\item \code{proficiency}: whether sample consisted of emergent or balanced bilinguals
\item \code{age}: mean age of the sample
\item \code{country}: country of study
\item \code{geo_area}: geographic area of study
\item \code{match}: did the study use matched samples (0 = no, 1 = yes)
\item \code{study_quality}: summated study quality score
\item \code{verbal_non_verbal}: whether task was verbal, non-verbal, or both
\item \code{outcome_task}: name of task used
\item \code{outcome_type}: coded for incongruent, congruent, and neutral trials
\item \code{sub_measure}: coded for reaction time, accuracy, or other outcomes
}
}
\usage{
data("lowe2020")
}
\description{
Results from 150 studies, including 1194 effect sizes (Hedge's g), on the extent to which
shared reading impacts language development \insertCite{lowe2020bilingual}{psymetadata}.
}
\details{
Further details can be found at \href{https://osf.io/jv7wt/}{https://osf.io/jv7wt/}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
