% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sala2019}
\alias{sala2019}
\title{Studies on the impact of working-memory training on near- and far-transfer measures}
\format{
A dataset with 1,555 rows and 15 variables.
* \code{id}: A numeric containing a unique id for each observed effect
* \code{study_id}: A numeric containing a unique id for each meta-analysis
* \code{study_name}: A character indicating the author, year, and comparison number (if applicable) of the meta-analysis
* \code{n}: A numeric containing the total sample size for the observed effect size in the meta-analysis
* \code{g}: A numeric containing the observed effect size, expressed in Hedge's g
* \code{se}: A numeric containing the standard error of the observed effect size
* \code{var}: A numeric containing the variance of the observed effect size
* \code{r}: A numeric containing the observed effect size, observed as r
* \code{comparison}:
* \code{age}:
* \code{type}:
* \code{test}:
* \code{model}:
}
\source{
\href{https://osf.io/qk2vu/}{Open Science Framework}
}
\usage{
data(sala2019)
}
\description{
To be filled in...

\insertNoCite{sala2019near}{psychmetadata}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
