% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spaniol2019}
\alias{spaniol2019}
\title{Studies on Executive function components in intellectual disability}
\format{
A data frame with 99 rows and 11 variables:
\itemize{
\item \code{author}: author of study
\item \code{study_id}: unique id for study
\item \code{study_year}: year of publication
\item \code{es_id}: unique id for effect size
\item \code{yi}: effect size in (Hedge's g)
\item \code{vi}: sampling variance for effect size (SE^2)
\item \code{group_id}: experimental intellectual disability group. one of: non-specific cause ("NSID"), Fragile X syndrome ("FXS"), Down syndrome ("DS"), or Williams syndrome ("WS")
\item \code{ef_type}: task type ("inhibition", "updating", "shifting", "fluency", "attention", or "other")
\item \code{ef_component}: executive function component ("inhibition", "shifting" or "attention")
\item \code{domain}: domain of executive function component ("verbal", "visuospatial", or "other")
\item \code{test}: test used to measure executive function
}
}
\usage{
data("spaniol2019")
}
\description{
Results from 26 studies, including 99 effect sizes (Hedge's g), on inhibition, shifting, and attention in people with intellectual disability compared to people matched on mental age \insertCite{spaniol2019meta}{psymetadata}.
}
\details{
Further details can be found at \href{https://osf.io/wp8u7/}{https://osf.io/wp8u7/}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
