% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stasielowicz2019b}
\alias{stasielowicz2019b}
\title{Studies on the association between goal orientation and performance adaptation}
\format{
A data frame with 86 rows and 25 variables:
\itemize{
\item \code{study_id}: unique id for study
\item \code{es_id}: unique id for effect size
\item \code{author}: author of study
\item \code{pub_year}: publication year
\item \code{pub_type}: publication type (0 = journal article, 2 = book chapter, 3 = dissertation, 4 = master's thesis, 5 = bachelor's thesis, 6 = conference proceedings, 7 = report, 8 = other)
\item \code{peer_review}: whethere publication was peer-reviewed (0 = no, 1 = yes)
\item \code{n}: sample size of effect size
\item \code{yi}: effect size (r)
\item \code{vi}: sampling variance of effect size (SE^2)
\item \code{pgo_type}: the performance goal orientation that was assessed ("avoid", "prove", or "global")
\item \code{adapt_measures}: assessment method(s) of adaptation used in the study (1 = self-report, 2 = other people, 3 = objective, 4 = mixed)
\item \code{adapt_method}: assessment method of adaption (0 = subjective ratings, 1 = objective scores)
\item \code{adapt_method_specific}: specific assessment method of adaptation used for the particular effect size (1 = self-report, 2 = other people, 3 = objective)
\item \code{go_measure}: instrument used to assess goal orientation
\item \code{financ_support}: financial support (0 = no, 1 = yes)
\item \code{age}: mean age of sample
\item \code{age_imputed}: mean age of sample (imputed)
\item \code{sex}: sex of sample (1 = female sample, 2 = male sample, 3 = mixed sample)
\item \code{perc_men}: proportion of men in the sample
\item \code{country}: country where sampled was collected
\item \code{sample}: sample type (1 = students, 2 = employees, 3 = manager, 4 = mixed, 5 = other)
\item \code{level}: level (1 = individuals, 2 = team)
\item \code{complexity_component}: component complexity of the task (0 = relatively low, 1 = relatively high)
\item \code{complexity_coordinative}: coordinative complexity of the task (0 = relatively low, 1 = relatively high)
\item \code{complexity_dynamic}: which complexity aspect changed while completing task (0 = neither component nor coordinative, 1 = only component, 2 = only coordinative, 3 = both component and coordinative)
}
}
\usage{
data("stasielowicz2019b")
}
\description{
Results from 28 studies, including 86 effect sizes (r), on performance goal orientation and performance adaptation \insertCite{stasielowicz2019goal}{psymetadata}.
}
\details{
Further details can be found at \href{https://osf.io/szfwx/}{https://osf.io/szfwx/}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
