% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stasielowicz2020}
\alias{stasielowicz2020}
\title{Studies on the importance of cognitive ability in performance adaptation}
\format{
A data frame with 133 rows and 23 variables:
\itemize{
\item \code{id}: unique id of study
\item \code{effect_id}: unique id of effect
\item \code{author}: author of study
\item \code{pub_year}: year of publication
\item \code{pub_type}: publication type (1 = journal article, 2 = book chapter, 3 = dissertation, 4 = master's thesis, 5 = bachelor's thesis, 6 = conference proceedings, 7 = report, 8 = other (eg., unpublished manuscript))
\item \code{peer_review}: whether publication was peer-reviewed (0 = no, 1 = yes)
\item \code{n}: sample size for effect size
\item \code{yi}: effect size (r)
\item \code{vi}: sampling variance (SE^2)
\item \code{adapt_measures}: assessment method(s) of adaptation used in the study (1 = self-report, 2 = other people, 3 = objective, 4 = mixed)
\item \code{adapt_method}: assessment method of adaption (1 = subjective ratings, 2 = objective scores)
\item \code{adapt_method_specific}: specific assessement method of adaptation used for the particular effect size (1 = self-report, 2 = other people, 3 = objective)
\item \code{subj_adapt_definition}: definition for subjective ratings of performance adaptations ("narrow" or "broad")
\item \code{cog_abil_measure}: measurement method of cognitive abilities
\item \code{ca_measure}: categorized measure of cognitive abilities ("general", "specific" or "ACT/SAT/GPA")
\item \code{financ_support}: financial support (e.g., grant; 0 = no, 1 = yes)
\item \code{sex}: sex (1 = female sample, 2 = male sample, 3 = mixed sample)
\item \code{men_prop}: proportion of men in sample
\item \code{country}: country of sample
\item \code{sample}: sample type (1 = students, 2 = employees, 3 = manager, 4 = mixed, 5 = other)
\item \code{task}: task used to measure performance adaptation ("simulation/video game", "SJT", or "Other")
\item \code{complexity_component}: coordinative complexity of the task (0 = relatively low, 1 = relatively high)
\item \code{complexity_coordinative}: dynamic complexity of the task (0 = relatively low, 1 = relatively high)
}
}
\usage{
data("stasielowicz2020")
}
\description{
Results from 47 independent samples, including 133 effect sizes (r),
on the role of individual differences in cognitive abilities in the context of performance adaption \insertCite{stasielowicz2020important}{psymetadata}.
}
\details{
Further details can be found at \href{https://psyarxiv.com/qu4t2/}{https://psyarxiv.com/qu4t2/}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
