% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stramaccia2020}
\alias{stramaccia2020}
\title{Studies on memory suppression and its deficiency in psychological disorders}
\format{
A data frame with 96 rows and 15 variables:
\itemize{
\item \code{study_id}: unique id for study
\item \code{group_id}: unique id for group
\item \code{es_id}: unique id for effect size
\item \code{yi}: effect size (Cohen's d)
\item \code{vi}: sampling variance for effect size (SE^2)
\item \code{pub_year}: year of publication
\item \code{instructions}: type of instructions given to participants to prevent retrieval ("aided", "direct", or "unspecified")
\item \code{stimuli}: type of stimuli ("pictures" or "words")
\item \code{valence}: valence of stimulus material (for the suppress targets only). One of "mixed", "negative", "neutral" or "positive"
\item \code{tnttime}: duration for which cues remained on the screen during the think/no-think phase (see paper for details)
\item \code{repetitions}: the number of times that participants encountered each cue in the think/no-think phase (see paper for details)
\item \code{n}: sample size
\item \code{dv}:
\item \code{cluster}: clusters based on clinical and sub-clinical conditions ("anxiety", "control", "depression", "mixed", or "repression")
\item \code{group}: clinical population ("CP") or healthy control ("HC")
}
}
\usage{
data("stramaccia2020")
}
\description{
Results from 25 studies, including 96 effect sizes (Cohen's d), on suppression-induced forgetting \insertCite{stramaccia2020memory}{psymetadata}.
}
\details{
Further details can be found at \href{https://osf.io/f89ur/}{https://osf.io/f89ur/}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
