% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nuijten2020}
\alias{nuijten2020}
\title{Data collected from meta-analyses on intelligence research}
\format{
A dataset with 2443 rows and 14 variables.
\itemize{
\item \code{study_id}: Unique id for study
\item \code{effect_id}: Unique id for effect size
\item \code{authors}: identifier for the primary study within a meta-analysis based o
the first author of the study or the sample used
\item \code{year}: year in which the primary study was reported
\item \code{yi}: original effect size converted to a Fishers z value
\item \code{vi}: variance around the z value in \code{yi}
\item \code{ni}: total sample size of the primary study
\item \code{es}: effect size as indicated in \code{type_es}
\item \code{se}: standard errors of the effect size
\item \code{type_es}: the type of effect size extracted from the meta-analysis; 1 = r transformed to Fishers z, 2 = Hedge's g, 3 = log odds ratio
4 = Cohen's d, 5 = Hazard Ratio, 6, 7, 8 = other
\item \code{type}: type of IQ research summarized by the meta-analysis; 1 = Correlational, 2 = Group_differences
3 = Experiments/Interventions, 4 = Toxicology, 5 = (Behavior) Genetics
\item \code{citations}: number of times the primary study was cited
\item \code{countrycode}: country in which the first author of a primary study was situated at the time of publication
\item \code{jrnl_impact}: impact factor in 2014 of the journal where the primary study was published
\item \code{similarity}: whether the primary study matched the research question of the meta-analysis; 0 = dissimilar, 1 = similar
}
}
\source{
\href{https://osf.io/fq5wp/}{https://osf.io/fq5wp/}
}
\usage{
data(nuijten2020)
}
\description{
Data resulting from 131 meta-analyses, including 2443 effect sizes (fisher-z), on different areas of intelligence research \insertCite{nuijten2020effect}{psymetadata}
}
\references{
\insertAllCited{}
}
\keyword{datasets}
